Replace old language server with new language server
