+++ effekt/jvm/src/main/scala/effekt/Driver.scala
@@ -67,7 +67,7 @@ trait Driver extends kiama.util.Compiler[EffektConfig, EffektError] { outer =>
         }
 
         // we are in one of three exclusive modes: LSPServer, Compile, Run
-        if (config.server() || config.experimentalServer()) { compiler.runFrontend(src) }
+        if (config.server()) { compiler.runFrontend(src) }
         else if (config.interpret()) { compile() foreach runner.eval }
         else if (config.build()) { compile() foreach runner.build }
         else if (config.compile()) { compile() }
